Synthesizer compile under windows

+++ Gammou/Process/abstract_component.h
@@ -199,7 +199,7 @@ namespace Gammou {
 		template<class T>
 		const unsigned int abstract_component<T>::get_input_count() const noexcept
 		{
-			return m_input_name.size();
+			return static_cast<unsigned int>(m_input_name.size());
 		}
 
 		template<class T>
